+++ doublex/__init__.py
@@ -1,7 +1,10 @@
 # -*- coding:utf-8; tab-width:4; mode:python -*-
 
+from functools import partial
+
 from .doubles import *
 from .matchers import *
+from .tracer import Tracer
 from .internal import WrongApiUsage